fixed bug

+++ road_segmentation_main/source/engine.py
@@ -437,7 +437,7 @@ class Engine:
         if add_penalty:
             kernel_size = Configuration.get("training.general.penalty.kernel_size", optional=True, default=3)
             lam = Configuration.get("training.general.penalty.lam", optional=True, default=0.5)
-            avgPool = torch.nn.AvgPool2d(kernel_size, stride=1, padding=np.floor(kernel_size/2))
+            avgPool = torch.nn.AvgPool2d(kernel_size, stride=1, padding=np.floor(kernel_size/2).astype(int))
             mean_predictions = avgPool(predictions)
             mse_loss = torch.nn.MSELoss()
             penalty = mse_loss(predictions, mean_predictions)
